src_lambdas/benefit_analysis/requirements.txt
pydantic==2.10.6
gradio==4.40.0
boto3==1.34.140
pillow==10.4.0
fastapi==0.111.0
mangum==0.17.0
pandas==2.2.2
uvicorn==0.23.2